feat: add wasm tx generator for snap sign (#4)
This commit introduces a WASM transaction parser and encoder to the
snap. The returned bytes of the `signTransaction` method will be a borsh
serialized transaction, that can be directly sent to sequencers.

It adds the bytecode of the wasm to the hardcoded snap source so its
source is committed to the hash of the published snap. It is done that
way to avoid exploits from the UI that might return malicious WASM files
and sign tampered messages for the user - having the bytes of the WASM
committed to the plugin asserts the signature will be only for the
intended message.